drm/msm: update msm gem and smmu for 4.14
This patch fixes some errors when merging downstream 4.9 code with
upstream 4.14 code and addresses issues with holding the global
struct_mutex locks as well as the increase in size of iova variables
from 32-bit to 64-bit values.  Some functions in msm_gem_vma.c were
moved around to improve readability.

+++ drivers/gpu/drm/msm/dsi-staging/dsi_ctrl.c
@@ -1390,7 +1390,7 @@ static int dsi_ctrl_buffer_deinit(struct dsi_ctrl *dsi_ctrl)
 int dsi_ctrl_buffer_init(struct dsi_ctrl *dsi_ctrl)
 {
 	int rc = 0;
-	u32 iova = 0;
+	u64 iova = 0;
 	struct msm_gem_address_space *aspace = NULL;
 
 	aspace = dsi_ctrl_get_aspace(dsi_ctrl, MSM_SMMU_DOMAIN_UNSECURE);
